requirements.txt
asgiref==3.2.9
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.7
djangorestframework==3.11.0
httpie==2.2.0
idna==2.10
mysqlclient==1.4.6
oauthlib==3.1.0
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9

django-allauth~=0.42.0
django-rest-auth~=0.9.5
pip~=19.2.3
click~=7.1.2
flex~=6.14.1
PyYAML~=5.3.1
Werkzeug~=1.0.1
jsonpointer~=2.0
rfc3987~=1.3.8
Jinja2~=2.11.2
MarkupSafe~=1.1.1
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
uritemplate~=3.0.1
inflection~=0.5.0
openapi~=1.1.0
packaging~=20.4
simplejson~=3.17.0
pyparsing~=2.4.7
setuptools~=41.2.0
Markdown~=3.2.2